requirements.txt
streamlit==1.31.1
torch==2.1.2
torchaudio==2.1.2
librosa==0.10.1
numpy==1.24.3
openai-whisper==20231117
speechbrain==0.5.16
pytube==15.0.0
plotly==5.18.0
pandas==2.1.4
Pillow==10.2.0
moviepy
transformers
ffmpeg-python
future
imageio
imageio-ffmpeg
pillow
python-dotenv
requests
scikit-learn
tqdm
PyYAML
filelock
huggingface-hub
regex
safetensors
